bash/functions: add s() function
This is a quick and dirty helper four my trouble with TERM
Kitty is awesome as a terminal-emulator but with TERM=xterm-kitty
every remote server shell will crap out b/c of unknown terminfo.
I simply use the shorthand s() to connect to remotes, which will
set TERM to 'xterm' in order to reduce anger and pain.